pyproject.toml
[project]
name = "crypto-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-binance>=1.0.25",
    "python-dotenv>=1.0.1",
]
